Update File_Creator.py
Changed the way the column headings are created.  Now, instead of column headings being the same as the categories selected in the GUI, the column headings match the fields in HMIS.  The categories in the GUI will remain the same as they are reflective of the term used in the SSVF Repository DQ reports.